query for the jobs that where updated after the last completion
CRAB jobs are currently hold in the queue after being completed, and they will be updated after they change to the status complete, we need to query if they were updated by the postjob after the last completion. 
Using the new CRAB_PostJobLastUpdate classad we can identify the latest modification.