Query the queues filtered the completed that hasn't changed
For the completed and removed jobs we are now filtering for the ones that have been updated in the last spider period. 

Assumptions: 
The CRAB_PostJobLastUpdate is updated by crab when the PostJobStatus is updated. 
The TIMEOUT_MINS constant is set to one minute less that the spider period. 

This change will keep sending the expected "snapshot" to MONiT.