Python3 migration (#137)
* Automatic conversion using 2to3 and black

```bash
2to3 -f all -f idioms -v -n -W . && black . 
```
Updating CMSMonitoring version

* Affiliation Manager bandit fixes

- Use a non public route for the affiliation dir
- Now we want to verify the ssl certificate. 

NOTE:  depending on the environment configuration it may not use the system's CAs, in those case you need to add this to the environment (centos 7) : 
```
 export REQUESTS_CA_BUNDLE=/etc/ssl/certs/ca-bundle.trust.crt
```

testDocConversion, pickle is binary (change the open mode for the files)

* P3 - bytes is not json serializable

* Autoflake

Remove unused imports

* fix some pylint Errors

* Add travis conf

* spider_osg - Remove a call to an nonexistent function.

* Add the CA bundle location to the Affiliation Manager cron

In order to verify the service certificate we need to be able to validate it against the CA certificate.

* TEMP - Use the Python3 branch for CMSMonitoring

* Base Dockerfile

* Use environment variables to set the monit flow environment

In order to use the test flow we need to be able to change of environment to the test one.

* Fix affiliation bug (20190727) which creates an empty file when the request fails.

* Fix affiliation bug (20190727) which creates an empty file when the request fails.

* fix future warning (split regex should not match empty)

* Manage an exception

If there are any exception in the processes we can have an empty queue at this point. We should treat this as a poison pill.

* add venv to .gitignore

* Replace the collector cmssrv221 with cmssrv623

* CMSTool in running fields

* Adding some fields to running jobs (committedTime and JobLastStartDate)

In order to show this values in the task monitoring details dashboard for the running jobs, we need to add them to the running_fields.

* Normalize PostJobStatus before decoding.

* Add itb collector to the default collectors.

* Set the postjobstatus for test tasks

* Add JobCurrentStartDate to running jobs

* Add pool name to the documents. 

This change makes the spider_volunteer script useless (as the collector is added as configuration to the cms spider)

* TaskType = CMS_TaskType if set, else fallback to guess it

* DO NOT MONIT

* Test jobs can also be analysis, check the jobtype.

* Housekepping

Deleting scripts that have been flagged as obsolete and redundant scripts that are not longed used.

* Update documentation and remove default collectors

Documentation about the collectors file and remove the default collectors.

* Capture ES Transport error

If this exception is not capture we can miss some documents in ES from other schedds that haven't been affected.